Create Binary Search
A binary search is an algorithm to find a particular element in the list !!! The elements in the list must be sorted to apply the binary search algorithm. If elements are not sorted then sort them first. We can find the element's index position very fast using the binary search algorithm.