Create Tic-Tac-Toe App.py
Here is an interesting tic-tac-toe game which all of us might have played in our childhood! You just have to enter a number according to the position as shown to make either X's or 0's  enter into the box. Faced a little difficulty in the starting of the code, but felt more and more interesting as it moved forward. I became a little curious to know how it would work !!! Enjoy playing :)