Create Codeforces problems.py
As I am new to codeforces, I have faced many difficulties in understanding the logic behind problems. But, it became interesting as I was continuing with it. I also felt sad as i did not get a few problems and others also, i got after getting errors. Though it was a little frustrating when i was not getting the solution, but i felt really happy when it used to show "accepted" !!! :)